+++ simple-chat/src/chat.html
@@ -20,10 +20,7 @@
             </div>
         
             <form id="message-form" class="message-form">
-                <textarea id="message-input" class="message-input" placeholder="Введите сообщение" rows="1"></textarea>
-                <input type="file" id="image-input" class="image-input" accept="image/*"">
-                <button type="button" id="image-button" class="image-button material-icons">image</button>
-                <button type="submit" class="send-button material-icons">send</button>
+                
             </form>
 
             <button id="delete-button" class="delete-button">
